# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InventoryItem {
  id: number;
  name: string;
  quantity: number;
  price: number;
  category: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateItemBody {
  name: string;
  quantity: number;
  price: number;
  category: string;
}

export interface UpdateItemBody {
  name?: string;
  quantity?: number;
  price?: number;
  category?: string;
}

export interface DeleteResult {
  success: boolean;
}

export type HistoryEventAction =
  (typeof HistoryEventAction)[keyof typeof HistoryEventAction];

export const HistoryEventAction = {
  created: "created",
  updated: "updated",
  deleted: "deleted",
  stock_added: "stock_added",
  stock_removed: "stock_removed",
} as const;

export interface HistoryEvent {
  id: number;
  itemId: number;
  itemName: string;
  action: HistoryEventAction;
  quantityBefore?: number | null;
  quantityAfter?: number | null;
  note?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalItems: number;
  totalStockValue: number;
  lowStockItems: InventoryItem[];
  totalCategories: number;
}

export type ListItemsParams = {
  /**
   * Search filter for item name or category
   */
  search?: string;
};

export type ListHistoryParams = {
  /**
   * Filter by item ID
   */
  itemId?: number;
};