# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all inventory items
 */
export const ListItemsQueryParams = zod.object({
  search: zod.coerce
    .string()
    .optional()
    .describe("Search filter for item name or category"),
});

export const ListItemsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  quantity: zod.number(),
  price: zod.number(),
  category: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListItemsResponse = zod.array(ListItemsResponseItem);

/**
 * @summary Create a new inventory item
 */
export const CreateItemBody = zod.object({
  name: zod.string(),
  quantity: zod.number(),
  price: zod.number(),
  category: zod.string(),
});

/**
 * @summary Get a single inventory item
 */
export const GetItemParams = zod.object({
  id: zod.coerce.number(),
});

export const GetItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  quantity: zod.number(),
  price: zod.number(),
  category: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update an inventory item
 */
export const UpdateItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateItemBody = zod.object({
  name: zod.string().optional(),
  quantity: zod.number().optional(),
  price: zod.number().optional(),
  category: zod.string().optional(),
});

export const UpdateItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  quantity: zod.number(),
  price: zod.number(),
  category: zod.string(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete an inventory item
 */
export const DeleteItemParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteItemResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List stock history events
 */
export const ListHistoryQueryParams = zod.object({
  itemId: zod.coerce.number().optional().describe("Filter by item ID"),
});

export const ListHistoryResponseItem = zod.object({
  id: zod.number(),
  itemId: zod.number(),
  itemName: zod.string(),
  action: zod.enum([
    "created",
    "updated",
    "deleted",
    "stock_added",
    "stock_removed",
  ]),
  quantityBefore: zod.number().nullish(),
  quantityAfter: zod.number().nullish(),
  note: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListHistoryResponse = zod.array(ListHistoryResponseItem);

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardResponse = zod.object({
  totalItems: zod.number(),
  totalStockValue: zod.number(),
  lowStockItems: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      quantity: zod.number(),
      price: zod.number(),
      category: zod.string(),
      createdAt: zod.date(),
      updatedAt: zod.date(),
    }),
  ),
  totalCategories: zod.number(),
});