# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateItemBody {
  name: string;
  quantity: number;
  price: number;
  category: string;
}