# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { InventoryItem } from "./inventoryItem";

export interface DashboardStats {
  totalItems: number;
  totalStockValue: number;
  lowStockItems: InventoryItem[];
  totalCategories: number;
}