# lib/api-zod/src/generated/types/historyEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEventAction } from "./historyEventAction";

export interface HistoryEvent {
  id: number;
  itemId: number;
  itemName: string;
  action: HistoryEventAction;
  quantityBefore?: number | null;
  quantityAfter?: number | null;
  note?: string | null;
  createdAt: Date;
}